adds ledger support for usb/browser
