+++ packages/neo-one-client/package.json
@@ -11,6 +11,8 @@
     "@neo-one/node-vm": "^1.0.0-alpha.69",
     "@neo-one/utils": "^1.0.0-alpha.57",
     "@reactivex/ix-esnext-esm": "^2.3.5",
+    "@ledgerhq/hw-transport-node-hid":"4.22.0",
+    "@ledgerhq/hw-transport-u2f":"4.21.0",
     "bignumber.js": "^7.2.1",
     "bn.js": "^4.11.8",
     "cross-fetch": "^2.2.2",
