+++ packages/neo-one-client/src/user/keystore/index.ts
@@ -1,3 +1,4 @@
-export * from './LocalKeyStore';
+export { LocalKeyStore } from './LocalKeyStore';
 export * from './LocalMemoryStore';
 export * from './LocalStringStore';
+export * from './TransportStore';
